added line to check for existing files before deletion

+++ voice_app/pages/1_Explore_Voice.py
@@ -194,16 +194,19 @@ def make_cnn_predictions(reshaped_data, scaler_path, model_path):
 
 
 def delete_temp_contents(temp_folder_path):
-    for file_name in os.listdir(temp_folder_path):
-        file_path = os.path.join(temp_folder_path, file_name)
-        try:
-            if os.path.isfile(file_path):
-                os.unlink(file_path)
-            elif os.path.isdir(file_path):
-                os.rmdir(file_path)
-        except Exception as e:
-            print(f"Error deleting {file_path}: {e}")
-
+    # Check if the folder is not empty
+    if any(os.scandir(temp_folder_path)):
+        for file_name in os.listdir(temp_folder_path):
+            file_path = os.path.join(temp_folder_path, file_name)
+            try:
+                if os.path.isfile(file_path):
+                    os.unlink(file_path)
+                elif os.path.isdir(file_path):
+                    os.rmdir(file_path)
+            except Exception as e:
+                print(f"Error deleting {file_path}: {e}")
+    else:
+        print(f"The folder {temp_folder_path} is already empty.")
 
 def main():
     # Get the absolute path to the current script
